requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
pydantic==1.10.8
pytest==7.3.1
requests==2.31.0
tomli==2.0.1
typing==3.7.4.3
typing_extensions==4.6.2
urllib3==2.0.2
